enum extensions and Application refactoring
