+++ clean-code-dotnetcore-api/src/Application/Application.csproj
@@ -14,6 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\CrossCutting.Automapper\CrossCutting.Automapper.csproj" />
+    <ProjectReference Include="..\CrossCutting.EnumExtensions\CrossCutting.EnumExtensions.csproj" />
     <ProjectReference Include="..\CrossCutting.FluentValidation\CrossCutting.FluentValidation.csproj" />
     <ProjectReference Include="..\Domain\Domain.csproj" />
   </ItemGroup>
