+++ clean-code-dotnetcore-api/src/Application/Queries/OrderWithProducts/OrderWithProductsQuery.cs
@@ -1,0 +1,11 @@
+﻿using Application.QueryModels;
+using MediatR;
+using System;
+
+namespace Application.Queries.OrderWithProducts
+{
+    public class OrderWithProductsQuery : IRequest<OrderWithProductsModel>
+    {
+        public Guid Id { get; set; }
+    }
+}
