+++ clean-code-dotnetcore-api/src/Application/Queries/Orders/OrdersQuery.cs
@@ -1,0 +1,11 @@
+﻿using Application.QueryModels;
+using MediatR;
+using System.Collections.Generic;
+
+namespace Application.Queries.Orders
+{
+    public class OrdersQuery : IRequest<List<OrderModel>>
+    {
+        // empty query
+    }
+}
