+++ clean-code-dotnetcore-api/src/Application/Queries/ProductCategories/ProductCategoriesQuery.cs
@@ -1,0 +1,11 @@
+﻿using Application.QueryModels;
+using MediatR;
+using System.Collections.Generic;
+
+namespace Application.Queries.ProductCategories
+{
+    public class ProductCategoriesQuery : IRequest<List<ProductCategoryModel>>
+    {
+        // empty query
+    }
+}
