+++ clean-code-dotnetcore-api/src/CrossCutting.Automapper/CrossCutting.Automapper.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CrossCutting.EnumExtensions\CrossCutting.EnumExtensions.csproj" />
+  </ItemGroup>
+
 </Project>
