+++ clean-code-dotnetcore-api/src/CrossCutting.EnumExtensions/CrossCutting.EnumExtensions.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+</Project>
