+++ clean-code-dotnetcore-api/src/CrossCutting.FluentValidation/CrossCutting.FluentValidation.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="3.1.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CrossCutting.EnumExtensions\CrossCutting.EnumExtensions.csproj" />
+  </ItemGroup>
+
 </Project>
